+++ src/app/(authorized)/@modal/(.)feed/create/page.tsx
@@ -1,10 +1,10 @@
-import { FeedCreateEditCard } from "@/components/feed/card";
+import { FeedCreateCard } from "@/components/feed/card";
 import { InterceptingModal } from "@/components/modal";
 
 export default function FeedCreateModal() {
   return (
     <InterceptingModal>
-      <FeedCreateEditCard />
+      <FeedCreateCard />
     </InterceptingModal>
   );
 }
