+++ src/app/(authorized)/@modal/(.)feed/edit/[id]/page.tsx
@@ -1,10 +1,10 @@
-import { FeedCreateEditCard } from "@/components/feed/card";
+import { FeedEditCard } from "@/components/feed/card";
 import { InterceptingModal } from "@/components/modal";
 
 export default function FeedEditModal() {
   return (
     <InterceptingModal>
-      <FeedCreateEditCard />
+      <FeedEditCard />
     </InterceptingModal>
   );
 }
