+++ src/app/(authorized)/feed/create/page.tsx
@@ -1,10 +1,10 @@
-import { FeedCreateEditCard } from "@/components/feed/card";
+import { FeedCreateCard } from "@/components/feed/card";
 import { ModalPage } from "@/components/modal";
 
 export default function FeedCreatePage() {
   return (
     <ModalPage>
-      <FeedCreateEditCard />
+      <FeedCreateCard />
     </ModalPage>
   );
 }
