+++ src/app/(authorized)/feed/edit/[id]/page.tsx
@@ -1,10 +1,10 @@
-import { FeedCreateEditCard } from "@/components/feed/card";
+import { FeedEditCard } from "@/components/feed/card";
 import { ModalPage } from "@/components/modal";
 
 export default function FeedEditPage() {
   return (
     <ModalPage type="feedDetail">
-      <FeedCreateEditCard />
+      <FeedEditCard />
     </ModalPage>
   );
 }
