+++ src/components/feed/card/index.tsx
@@ -1,2 +1,3 @@
 export { default as FeedDetailCard } from "./FeedDetailCard";
-export { default as FeedCreateEditCard } from "./FeedCreateEditCard";
+export { default as FeedCreateCard } from "./FeedCreateCard";
+export { default as FeedEditCard } from "./FeedEditCard";
